# Use pure Go based ZSTD implementation