--- a/go.mod
+++ b/go.mod
@@ -3,13 +3,14 @@ module github.com/dgraph-io/badger/v2
 go 1.12
 
 require (
-	github.com/DataDog/zstd v1.4.1
 	github.com/cespare/xxhash v1.1.0
 	github.com/dgraph-io/ristretto v0.0.0-20191025175511-c1f00be0418e
 	github.com/dgryski/go-farm v0.0.0-20190423205320-6a90982ecee2
 	github.com/dustin/go-humanize v1.0.0
 	github.com/golang/protobuf v1.3.1
 	github.com/golang/snappy v0.0.1
+	github.com/google/go-cmp v0.3.1 // indirect
+	github.com/klauspost/compress v1.9.4
 	github.com/kr/pretty v0.1.0 // indirect
 	github.com/pkg/errors v0.8.1
 	github.com/spaolacci/murmur3 v1.1.0 // indirect