--- a/table/builder.go
+++ b/table/builder.go
@@ -347,7 +347,7 @@ func (b *Builder) compressData(data []byte) ([]byte, error) {
 	case options.Snappy:
 		return snappy.Encode(nil, data), nil
 	case options.ZSTD:
-		return y.ZSTDCompress(nil, data, b.opt.ZSTDCompressionLevel)
+		return y.ZSTDCompress(nil, data)
 	}
 	return nil, errors.New("Unsupported compression type")
 }