--- a/table/table.go
+++ b/table/table.go
@@ -69,9 +69,6 @@ type Options struct {
 	Compression options.CompressionType
 
 	Cache *ristretto.Cache
-
-	// ZSTDCompressionLevel is the ZSTD compression level used for compressing blocks.
-	ZSTDCompressionLevel int
 }
 
 // TableInterface is useful for testing.