Fixes https://github.com/dgraph-io/badger/issues/1162

This PR proposes to use https://github.com/klauspost/compress/tree/master/zstd instead of CGO based https://github.com/DataDog/zstd . 

This PR also removes the `CompressionLevel` options since https://github.com/klauspost/compress/tree/master/zstd supports only two levels of ZSTD Compression. The default level is `ZSTD Level 3` and the `fastest` level is `ZSTD level 1`. `ZSTD level 1` will be the default level in badger.

I've experimented will all the suggestions mentioned in https://github.com/klauspost/compress/issues/196#issuecomment-568905095 . Setting `WithSingleSegment` didn't seem to make a lot of speed difference (~ 1MB/s difference)
`WithNoEntropyCompression` seemed to have ~ 3% of speed improvement (but that could also be because of non-deterministic nature of benchmarks)
```
name                                       old time/op      new time/op (NoEntropy set)   delta
Compression/ZSTD_-_Go_-_level1-16           35.7µs ± 1%     36.9µs ± 5%                 +3.41%  (p=0.008 n=5+5)
Decompression/ZSTD_-_Go-16                  16.0µs ± 0%     15.9µs ± 1%                 -0.77%  (p=0.016 n=5+5)

name                                    old speed      new speed (NoEntropy set)      delta
Compression/ZSTD_-_Go_-_level1-16      115MB/s ± 1%   111MB/s ± 5%                -3.24%  (p=0.008 n=5+5)
Decompression/ZSTD_-_Go-16             256MB/s ± 0%   258MB/s ± 1%                 +0.78%  (p=0.016 n=5+5)
```

### Benchmarks
1. Table Data (contains some randomly generated data). 
```
Compression Ratio Datadog ZSTD level 1 3.1993720565149135
Compression Ratio Datadog ZSTD level 3 3.099619771863118

Compression Ratio Go ZSTD 3.2170481452249406
Compression Ratio Go ZSTD level 3 3.1474903474903475
```
```
name                                        time/op
Compression/ZSTD_-_Datadog-level1-16    17.6µs ± 3%
Compression/ZSTD_-_Datadog-level3-16    20.7µs ± 3%

Compression/ZSTD_-_Go_-_level1-16       27.8µs ± 2%
Compression/ZSTD_-_Go_-_Default-16      39.1µs ± 1%

Decompression/ZSTD_-_Datadog-16         7.12µs ± 2%
Decompression/ZSTD_-_Go-16              13.7µs ± 2%

name                                       speed
Compression/ZSTD_-_Datadog-level1-16   231MB/s ± 3%
Compression/ZSTD_-_Datadog-level3-16   197MB/s ± 3%

Compression/ZSTD_-_Go_-_level1-16      147MB/s ± 2%
Compression/ZSTD_-_Go_-_Default-16     104MB/s ± 1%

Decompression/ZSTD_-_Datadog-16        573MB/s ± 2%
Decompression/ZSTD_-_Go-16             298MB/s ± 2%
```

2. 4KB of text taken from https://gist.github.com/StevenClontz/4445774
```
Compression Ratio ZSTD level 1 1.9294781382228492
Compression Ratio ZSTD level 3 1.9322033898305084

Compression Ratio Go ZSTD 1.894736842105263
Compression Ratio Go ZSTD level 3 1.927665570690465
```
```
name                                       time/op
Compression/ZSTD_-_Datadog-level1-16    22.7µs ± 4%
Compression/ZSTD_-_Datadog-level3-16    29.6µs ± 4%

Compression/ZSTD_-_Go_-_level1-16       35.7µs ± 1%
Compression/ZSTD_-_Go_-_Default-16      97.9µs ± 1%

Decompression/ZSTD_-_Datadog-16         8.36µs ± 0%
Decompression/ZSTD_-_Go-16              16.0µs ± 0%

name                                       speed
Compression/ZSTD_-_Datadog-level1-16   181MB/s ± 4%
Compression/ZSTD_-_Datadog-level3-16   139MB/s ± 4%

Compression/ZSTD_-_Go_-_level1-16      115MB/s ± 1%
Compression/ZSTD_-_Go_-_Default-16    41.9MB/s ± 1%

Decompression/ZSTD_-_Datadog-16        489MB/s ± 2%
Decompression/ZSTD_-_Go-16             256MB/s ± 0%
```

Here's the script I've used https://gist.github.com/jarifibrahim/91920e93d1ecac3006b269e0c05d6a24

<!-- Reviewable:start -->
---
This change is [<img src="https://reviewable.io/review_button.svg" height="34" align="absmiddle" alt="Reviewable"/>](https://reviewable.io/reviews/dgraph-io/badger/1176)
<!-- Reviewable:end -->