@jarifibrahim 2020-01-13
I had a chat with @manishrjain and we've decided to not use the `pure go ZSTD` because it's about 1.5x slower than the CGO based implementation.

```
Compression/ZSTD_-_Datadog-level1-16    22.7µs ± 4%
Compression/ZSTD_-_Go_-_level1-16       35.7µs ± 1%

Compression/ZSTD_-_Datadog-level3-16    29.6µs ± 4%
Compression/ZSTD_-_Go_-_Default-level3-16      97.9µs ± 1%

Decompression/ZSTD_-_Datadog-16         8.36µs ± 0%
Decompression/ZSTD_-_Go-16              16.0µs ± 0%
```