@klauspost 2020-06-02
@jarifibrahim Reran your script.

```
BenchmarkComp/Compression/ZSTD_-_Datadog-32        31495             37848 ns/op         107.70 MB/s
BenchmarkComp/Compression/ZSTD_-_Go_-_Fastest-32                   49791             23325 ns/op         174.75 MB/s
BenchmarkComp/Compression/ZSTD_-_Go_-_Default-32                   30686             36927 ns/op         110.38 MB/s
BenchmarkComp/Decompression/ZSTD_-_Datadog-32                     166665              7074 ns/op         576.18 MB/s
BenchmarkComp/Decompression/ZSTD_-_Go-32                          109090             10542 ns/op         386.65 MB/s
```
So compression both in 'fast' and 'default' is faster than cgo default.

"1.5x slower" doesn't make sense to me. If you mean that runtime is 150% of cgo, that does seem to be the case for decompression on this particular payload.

This has a very skewed performance profile since blocks are so small. I can look into the 'very small block' decompression performance.