@jarifibrahim 2020-06-02
Hey @klauspost, the `1.5x`  number (which is technically 2x in my original comment) was for the decompression speed. In badger, we **compress a block once but we have to decompress is multiple times**. Slow compression speed is okay since we do compression in the background but slow decompression speed would mean the reads become slower.

@klauspost we'd love to use the pure go based implementation if we can improve the decompression speed. I don't have experience with compression algorithms but if there's anyway I can help, please do let me know.

Thank for looking into this issue :)