@klauspost 2020-06-02
@jarifibrahim It depends so much on how you run the benchmark and a single (or two) payloads can skew the numbers significantly.

For example, look at these numbers:

```
BenchmarkComp/Decompression/ZSTD_-_Datadog-32             472735              24
75 ns/op        1646.84 MB/s
BenchmarkComp/Decompression/ZSTD_-_Go-32                 1124998              10
14 ns/op        4018.83 MB/s
```

These are the numbers by simply running a parallel benchmark:

```Go
		b.Run("ZSTD - Datadog", func(b *testing.B) {
			b.SetBytes(int64(len(data)))
			b.RunParallel(func(pb *testing.PB) {
				buf := make([]byte, len(data))
				for pb.Next() {
					d, err := zstd.Decompress(buf, ZSTDCompressed)
					if err != nil {
						panic(err)
					}
					_ = d
				}
			})
		})
		b.Run("ZSTD - Go", func(b *testing.B) {
			b.SetBytes(int64(len(data)))
			dec, err := gozstd.NewReader(nil)
			if err != nil {
				panic(err)
			}
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				buf := make([]byte, len(data))
				for pb.Next() {
					d, err := dec.DecodeAll(ZSTDCompressed, buf[:0])
					if err != nil {
						panic(err)
					}
					_ = d
				}
			})
		})
```

With this small change the Go version beats decompression speed of the CGO version by more than 2x. I would say this leads to a quite different conclusion AND it is a bit closer to what you would see in the real world.