@jarifibrahim 2020-06-02
@klauspost wow, I did not anticipate that. Why does the parallel version work so much faster? Or why is the CGO one significantly slower?

We have compression/decompression running parallelly all the time and from this new benchmark, it looks like the go based implementation would be very efficient. This is awesome.
Quick question - how big was your data for this decompression benchmark? 4 KB? Can you share your benchmark script?