@klauspost 2020-06-02
TBH I am a bit surprised myself ;)  My guess is that the cgo version allocates new memory on every run and trashes the cache.

It is your script linked above with only the lines above.