@klauspost 2020-06-02
The Go version will only allocate GOMAXPROCS decompressors and re-use them across goroutines thus limiting the total amount of memory used. The cgo version just allocates on every run.

I suspect actual performance is somewhere in between and not as extreme as seen above since other stuff is going on between compression runs.