@jarifibrahim 2020-06-02
@klauspost why are my results so different than yours?
```
 go test -run xxx -bench Comp/Decompression        
data size 4096
goos: linux
goarch: amd64
pkg: github.com/dgraph-io/badger/v2/table
BenchmarkComp/Decompression/Snappy-8  	  756020	      1607 ns/op	2549.00 MB/s
BenchmarkComp/Decompression/LZ4-8     	 1399564	       862 ns/op	4749.56 MB/s
BenchmarkComp/Decompression/ZSTD_-_Datadog-8         	  426633	      2859 ns/op	1432.49 MB/s
BenchmarkComp/Decompression/ZSTD_-_Go-8              	  194971	      6082 ns/op	 673.50 MB/s
PASS
ok  	github.com/dgraph-io/badger/v2/table	7.432s

```
```go
	b.Run("Decompression", func(b *testing.B) {
		buf := make([]byte, len(data))
		b.Run("Snappy", func(b *testing.B) {
			b.SetBytes(int64(len(data)))
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				for pb.Next() {
					d, err := snappy.Decode(buf, snappyCompressed)
					if err != nil {
						panic(err)
					}
					_ = d
					if validate {
						require.Equal(b, d, data)
					}
				}
			})
		})
		b.Run("LZ4", func(b *testing.B) {
			b.SetBytes(int64(len(data)))
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				for pb.Next() {
					n, err := lz4.UncompressBlock(LZ4Compressed, buf)
					if err != nil {
						fmt.Println(err)
					}
					buf = buf[:n] // uncompressed data
					if validate {
						require.Equal(b, buf, data)
					}
				}
			})
		})
		b.Run("ZSTD - Datadog", func(b *testing.B) {
			b.SetBytes(int64(len(data)))
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				for pb.Next() {
					d, err := zstd.Decompress(buf, ZSTDCompressed)
					if err != nil {
						panic(err)
					}
					_ = d
					if validate {
						require.Equal(b, d, data)
					}
				}
			})
		})
		b.Run("ZSTD - Go", func(b *testing.B) {
			b.SetBytes(int64(len(data)))
			dec, err := gozstd.NewReader(nil)
			if err != nil {
				panic(err)
			}
			b.ResetTimer()
			b.RunParallel(func(pb *testing.PB) {
				for pb.Next() {
					d, err := dec.DecodeAll(ZSTDCompressed, buf[:0])
					if err != nil {
						panic(err)
					}
					_ = d
					if validate {
						require.Equal(b, d, data)
					}
				}
			})
		})
```