@klauspost 2020-06-02
You are sharing the output buffer between goroutines.

Also, are you using `v1.10.7`?