@klauspost 2020-06-02
@jaipradeesh Other than that, the core count 4 vs. 16 is probably making a difference. The cgo version seems to saturate at a low concurrency level.