@jarifibrahim 2020-06-02
>You are sharing the output buffer between goroutines.

Yeah. Fixed that.

>Also, are you using v1.10.7?

Yes, I'm using v1.10.7

>@jarifibrahim Other than that, the core count 4 vs. 16 is probably making a difference. The cgo version seems to saturate at a low concurrency level.

That could be the reason. Let me try this on a different machine and get back.