@klauspost 2020-06-02
Cool. I'm using the generated table, btw.